requirements.txt
pyyaml>=6.0
requests>=2.31.0
openai>=1.0.0
tqdm>=4.66.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
